Move exception provider to custom command triggered by context menu. Also improved suggestion flow
